backend/requirements.txt
appdirs==1.4.4
python-dotenv
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
chromedriver-py==137.0.7151.70
cryptography==36.0.2
Django==3.2.8
django-cors-headers==3.11.0
django-crontab==0.7.1
django-extensions==3.1.5
django-rest-auth==0.9.5
djangorestframework==3.13.1
docutils==0.18.1
h11==0.13.0
idna==3.3
lockfile==0.12.2
lxml==4.8.0
memory-profiler==0.60.0
numpy==1.22.3
outcome==1.1.0
pandas==1.4.1
passlib==1.7.4
pid==3.0.4
psutil==5.9.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
selenium==4.1.3
service==0.6.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.63.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
webdrivermanager==0.10.0
wsproto==1.1.0
